requirements.txt
aiohttp[speedups]
orjson
appdirs
django
beautifulsoup4
lxml
requests
numpy
pandas
TA-Lib
croniter
asynctest
django-choices
mysqlclient
tqdm
pytz
tzdata
einops
huggingface_hub
matplotlib
safetensors
sshtunnel
pybind11
PyYAML